src/slices/recurrence/controllers/updateRecurrence/updateRecurrenceController.ts
/* eslint-disable no-unsafe-optional-chaining */
import {
  HttpRequest,
  HttpResponse,
  Validation,
  badRequest,
  ok,
} from "@/application/helpers";
import { Controller } from "@/application/infra/contracts";
import { UpdateRecurrence } from "@/slices/recurrence/useCases";

export class UpdateRecurrenceController extends Controller {
  constructor(
    private readonly validationQuery: Validation,
    private readonly validationBody: Validation,
    private readonly updateRecurrence: UpdateRecurrence
  ) {
    super();
  }
  async execute(httpRequest: HttpRequest<any>): Promise<HttpResponse<any>> {
    const errorsBody = this.validationBody.validate(httpRequest?.body);
    if (errorsBody?.length > 0) {
      return badRequest(errorsBody);
    }
    const errorsQuery = this.validationQuery.validate(httpRequest?.query);
    if (errorsQuery?.length > 0) {
      return badRequest(errorsQuery);
    }
    const recurrenceUpdated = await this.updateRecurrence(
      {
        ...httpRequest?.query,
        createdById: httpRequest?.userId,
      },
      httpRequest?.body
    );
    return ok(recurrenceUpdated);
  }
}
